backend/requirements.txt
Flask
flask-cors
python-docx
PyMuPDF
PyPDF2
torch
transformers
nltk
